server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pradyun_chat"
version = "1.0.0"
description = "Coding Assignment that interacts with Claude and OpenAI"
authors = [
    {name = "Pradyun Magal"}
]
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "flask",
    "openai",
    "anthropic",
    "pydantic",
    "python-dotenv",
    "flask-cors"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy"
]

[tool.setuptools]
package-dir = {"" = "src"}
